requirements.txt
cv2
pytesseract
os
PIL
sys
numpy
selenium
selenium.webdriver.common.action_chains
webdriver_manager.chrome
pandas
bs4
json
requests
time
os
urllib.request
random
selenium.webdriver.firefox.firefox_binary
selenium.webdriver.common.desired_capabilities
datetime